fix(routines,cli): split files back under the 500-line pre-push gate
src/routines/service_trigger.rs (505 lines) and src/cli.rs (508 lines)
each independently passed their own PR's CI, but two separate PRs
landing on main pushed both combined over the repo's 500-line-per-file
convention (linecheck isn't in the branch ruleset's required checks, so
neither PR's merge was blocked by it) — main's local pre-push hook now
fails this gate for anyone trying to push, release included.

Extracts svc_run_log/svc_run_summary (and their shared svc_run_file
helper) from service_trigger.rs into a new service_run_files.rs sibling,
and restart/restart_rotation_line/restart_json/start_detached_and_report/
report_endpoints from cli.rs into a new cli_restart.rs sibling — mirroring
the repo's existing cli_query.rs/service_trigger_flags.rs split pattern.
No behavior change.
