+++ src/routines/service.rs
@@ -406,9 +406,13 @@ use service_trigger::migrate_workbenches;
 #[cfg(test)]
 pub(crate) use service_trigger::sh_bin;
 pub use service_trigger::{
-    svc_cleanup, svc_list_all_runs, svc_list_runs, svc_logs, svc_run_log, svc_run_summary,
-    svc_set_power_saving, svc_snooze, svc_trigger, svc_trigger_scheduled,
+    svc_cleanup, svc_list_all_runs, svc_list_runs, svc_logs, svc_set_power_saving, svc_snooze,
+    svc_trigger, svc_trigger_scheduled,
 };
+
+#[path = "service_run_files.rs"]
+mod service_run_files;
+pub use service_run_files::{svc_run_log, svc_run_summary};
 
 #[path = "service_trigger_flags.rs"]
 mod service_trigger_flags;
